+++ models/db/UserNotification.php
@@ -19,9 +19,10 @@ use yii\db\ActiveRecord;
  */
 class UserNotification extends ActiveRecord
 {
-    const NOTIFICATION_NEW_MOTION    = 0;
-    const NOTIFICATION_NEW_AMENDMENT = 1;
-    const NOTIFICATION_NEW_COMMENT   = 2;
+    const NOTIFICATION_NEW_MOTION          = 0;
+    const NOTIFICATION_NEW_AMENDMENT       = 1;
+    const NOTIFICATION_NEW_COMMENT         = 2;
+    const NOTIFICATION_AMENDMENT_MY_MOTION = 3;
 
     /**
      * @return string
@@ -59,6 +60,26 @@ class UserNotification extends ActiveRecord
             [['userId', 'consultationId', 'notificationType'], 'required'],
             [['id', 'userId', 'consultationId', 'notificationType', 'notificationReferenceId'], 'number'],
         ];
+    }
+
+    /**
+     * @param Consultation $consultation
+     * @param null|int $notiType
+     * @return UserNotification[]
+     */
+    public static function getConsultationNotifications(Consultation $consultation, $notiType = null)
+    {
+        if ($notiType) {
+            $notifications = [];
+            foreach ($consultation->userNotifications as $userNotification) {
+                if ($userNotification->notificationType == $notiType) {
+                    $notifications[] = $userNotification;
+                }
+            }
+            return $notifications;
+        } else {
+            return $consultation->userNotifications;
+        }
     }
 
     /**