+++ History.md
@@ -3,6 +3,7 @@
 ## Version 3.7.0 [Not yet released]
 
 - Initiators of motions can now merge amendments into their motions, if this is allowed by the consultation settings. This function can be restricted to cases where no amendment rewriting is necessary.
+- Proposers of motions now can receive an e-mail if a amendment to their motion is published (enabled by default)
 - We don't pretend anymore you could use Antragsgrün with IE <=9
 
 ### Version 3.6.2 [Not yet released]